core/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "copilotkit>=0.1.87",
    "fastapi>=0.136.0",
    "google-adk>=1.10.0",
    "openai>=2.32.0",
    "psutil>=7.2.2",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.11",
    "tree-sitter>=0.25.2",
    "tree-sitter-css>=0.25.0",
    "tree-sitter-html>=0.23.2",
    "tree-sitter-markdown>=0.5.1",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
    "uvicorn>=0.44.0",
    "websockets>=16.0",
]
[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::pytest.PytestRemovedIn9Warning",
]

